requirements.txt
pymongo
python-dateutil
requests
xmltodict
lxml